CTW-801/remove 1.0 components (#530)
Remove old conditions & medications components in favor of new design
with tabs for patient records and other provider records.

Renamed `PatientMedicationsTabbed` to `PatientMedications` as it is now
the default view for medications.
