+++ .changeset/seven-maps-repeat.md
@@ -1,0 +1,5 @@
+---
+"@zus-health/ctw-component-library": minor
+---
+
+Remove old conditions & medications components in favor of new design with tabs for patient records and other provider records.
