+++ src/components/content/condition-history/conditions-history.tsx
@@ -7,7 +7,7 @@ import {
 import { useCTW } from "../../core/providers/ctw-provider";
 import { DocumentButton } from "../CCDA/document-button";
 import { useCCDAModal } from "../CCDA/modal-ccda";
-import { ConditionHeader } from "../condition-header";
+import { ConditionHeader } from "../conditions/condition-header";
 import { applyConditionHistoryFilters } from "./condition-history-filters";
 import { conditionData, setupData } from "./condition-history-schema";
 import { withErrorBoundary } from "@/components/core/error-boundary";
