+++ src/components/content/medications/patient-medications.stories.tsx
@@ -1,14 +1,11 @@
-import { expect } from "@storybook/jest";
 import type { Meta, StoryObj } from "@storybook/react";
 import { setupMedicationMocks } from "./story-helpers/mocks/requests";
 import {
   PatientMedications,
   PatientMedicationsProps,
 } from "@/components/content/medications/patient-medications";
-import { medicationFormDrawer } from "@/components/content/medications/story-helpers/medication-form-drawer";
 import { otherProviderMedications } from "@/components/content/medications/story-helpers/mocks/other-provider-medications";
 import { providerMedications } from "@/components/content/medications/story-helpers/mocks/provider-medications";
-import { medicationsTables } from "@/components/content/medications/story-helpers/patient-medications";
 import { CTWProvider } from "@/components/core/providers/ctw-provider";
 import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
@@ -27,74 +24,28 @@ export default {
       </CTWProvider>
     ),
   ],
+  args: {
+    hideAddToRecord: undefined,
+    forceHorizontalTabs: undefined,
+    onAfterOpenHistoryDrawer: undefined,
+    onOpenHistoryDrawer: undefined,
+  },
 } as Meta<Props>;
 
 export const Basic: StoryObj<Props> = {
   ...setupMedicationMocks({ providerMedications, otherProviderMedications }),
 };
 
+export const ForceHorizontalTabs: StoryObj<Props> = {
+  ...Basic,
+  args: {
+    forceHorizontalTabs: true,
+  },
+};
+
 export const HideAddToRecord: StoryObj<Props> = {
   ...Basic,
   args: {
     hideAddToRecord: true,
-  },
-};
-
-export const TestAddNewMed: StoryObj<Props> = {
-  ...Basic,
-  play: async ({ canvasElement }) => {
-    const medications = await medicationsTables(canvasElement);
-    await medications.patientRecord.toHaveRowCount(1);
-    const newMedication = "albendazole 200 MG Oral Tablet [Albenza]";
-    medications.clickAddMedication();
-    const addMedicationForm = await medicationFormDrawer(canvasElement);
-    addMedicationForm.search("alb");
-    await addMedicationForm.selectMedication(newMedication);
-    addMedicationForm.status("Active");
-    addMedicationForm.instructions(
-      "Take 2.3 every 3.14 hours, if dose is missed spin around 4.5 times"
-    );
-    await addMedicationForm.save();
-    await medications.patientRecord.toHaveRowCount(2);
-    expect(
-      await medications.patientRecord.table.findByText(newMedication)
-    ).toBeTruthy();
-  },
-};
-
-export const TestAddToRecord: StoryObj<Props> = {
-  ...Basic,
-  play: async ({ canvasElement }) => {
-    const medications = await medicationsTables(canvasElement);
-    await medications.patientRecord.toHaveRowCount(1);
-    await medications.otherProvider.toHaveRowCount(4);
-    const medicationName =
-      "3 ML insulin aspart protamine, human 70 UNT/ML / insulin aspart, human 30 UNT/ML Pen Injector [NovoLog Mix]";
-    medications.otherProvider.toHaveRowWithText(0, medicationName);
-    await medications.otherProvider.addToRecord(0);
-    const medDrawer = await medicationFormDrawer(canvasElement);
-    await medDrawer.save();
-    await medications.patientRecord.toHaveRowCount(2);
-    await medications.otherProvider.toHaveRowCount(3);
-    await medications.patientRecord.toHaveAnyRowWithText(medicationName);
-  },
-};
-
-export const TestCancelAddNewMed: StoryObj<Props> = {
-  ...Basic,
-  play: async ({ canvasElement }) => {
-    const medications = await medicationsTables(canvasElement);
-    await medications.patientRecord.toHaveRowCount(1);
-    const newMedication = "cabozantinib 20 MG Oral Capsule [Cometriq]";
-    medications.clickAddMedication();
-    const addMedicationForm = await medicationFormDrawer(canvasElement);
-    addMedicationForm.search("cab");
-    await addMedicationForm.selectMedication(newMedication);
-    addMedicationForm.status("Active");
-    addMedicationForm.cancel();
-    await medications.patientRecord.toHaveRowCount(1);
-    expect(
-      medications.patientRecord.table.queryByText(newMedication)
-    ).toBeFalsy();
   },
 };
